+++ litebird_sim/mapmaking/binner.py
@@ -68,7 +68,7 @@ class BinnerResult:
 
 @njit
 def _solve_binning(nobs_matrix, atd):
-    # Sove the map-making equation
+    # Solve the map-making equation
     #
     # This method alters the parameter `nobs_matrix`, so that after its completion
     # each 3×3 matrix in nobs_matrix[idx, :, :] will be the *inverse*.
@@ -224,6 +224,7 @@ def _build_nobs_matrix(
         pixidx_all, polang_all = _compute_pixel_indices(
             hpx=hpx,
             pointings=cur_ptg,
+            pol_angle_detectors=cur_obs.pol_angle_rad,
             num_of_detectors=cur_obs.n_detectors,
             num_of_samples=cur_obs.n_samples,
             hwp_angle=hwp_angle,